+++ receiver/prometheusreceiver/internal/otlp_metrics_adjuster.go
@@ -21,6 +21,7 @@ import (
 	"time"
 
 	"go.opentelemetry.io/collector/model/pdata"
+
 	"go.uber.org/zap"
 )
